requirements.txt
django
gunicorn
django-heroku
django-single-instance-model
selenium